add tests for new modules
