+++ tests/CMakeLists.txt
@@ -21,6 +21,9 @@ list(APPEND target_tests "ProgressBarTest")
 list(APPEND target_tests "CsvIOTest")
 list(APPEND target_tests "H5IOTest")
 list(APPEND target_tests "YAMLGetFieldTest")
+list(APPEND target_tests "DefEnumTest")
+list(APPEND target_tests "DefTupleTest")
+list(APPEND target_tests "CsvIOTest")
 
 if ("CsvIOTest" IN_LIST target_tests)
     set(test_name "CsvIOTest.generic")
@@ -45,6 +48,20 @@ if ("H5IOTest" IN_LIST target_tests)
 endif()
 if ("NumericTimeTest" IN_LIST target_tests)
     set(test_name "NumericTimeTest")
+    add_executable(${test_name})
+    target_sources(${test_name} PUBLIC ${test_name}.cpp)
+    target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain WCCommon::WCCommon)
+    add_test("${test_name}" ${test_name})
+endif()
+if ("DefEnumTest" IN_LIST target_tests)
+    set(test_name "DefEnumTest")
+    add_executable(${test_name})
+    target_sources(${test_name} PUBLIC ${test_name}.cpp)
+    target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain WCCommon::WCCommon)
+    add_test("${test_name}" ${test_name})
+endif()
+if ("DefTupleTest" IN_LIST target_tests)
+    set(test_name "DefTupleTest")
     add_executable(${test_name})
     target_sources(${test_name} PUBLIC ${test_name}.cpp)
     target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain WCCommon::WCCommon)
